+++ platform/darwin/src/NSExpression+MGLAdditions.mm
@@ -656,8 +656,12 @@ NS_DICTIONARY_OF(NSNumber *, NSExpression *) *MGLStopDictionaryByReplacingTokens
     return [NSExpression expressionForVariable:@"featureIdentifier"];
 }
 
++ (NSExpression *)featureAttributesVariableExpression {
+    return [NSExpression expressionForVariable:@"featureAttributes"];
+}
+
 + (NSExpression *)featurePropertiesVariableExpression {
-    return [NSExpression expressionForVariable:@"featureProperties"];
+    return [self featureAttributesVariableExpression];
 }
 
 + (instancetype)mgl_expressionForConditional:(nonnull NSPredicate *)conditionPredicate trueExpression:(nonnull NSExpression *)trueExpression falseExpresssion:(nonnull NSExpression *)falseExpression {
@@ -901,7 +905,7 @@ NSArray *MGLSubexpressionsWithJSONObjects(NSArray *objects) {
         } else if ([op isEqualToString:@"id"]) {
             return NSExpression.featureIdentifierVariableExpression;
         }  else if ([op isEqualToString:@"properties"]) {
-            return NSExpression.featurePropertiesVariableExpression;
+            return NSExpression.featureAttributesVariableExpression;
         } else if ([op isEqualToString:@"var"]) {
             return [NSExpression expressionForVariable:argumentObjects.firstObject];
         } else if ([op isEqualToString:@"case"]) {
@@ -1000,7 +1004,7 @@ NSArray *MGLSubexpressionsWithJSONObjects(NSArray *objects) {
             if ([self.variable isEqualToString:@"featureIdentifier"]) {
                 return @[@"id"];
             }
-            if ([self.variable isEqualToString:@"featureProperties"]) {
+            if ([self.variable isEqualToString:@"featureAttributes"]) {
                 return @[@"properties"];
             }
             return @[@"var", self.variable];