+++ platform/darwin/test/MGLExpressionTests.mm
@@ -183,9 +183,9 @@ using namespace std::string_literals;
         XCTAssertEqualObjects([NSExpression expressionWithMGLJSONObject:@[@"id"]], expression);
     }
     {
-        NSExpression *expression = [NSExpression expressionForVariable:@"featureProperties"];
+        NSExpression *expression = [NSExpression expressionForVariable:@"featureAttributes"];
         XCTAssertEqualObjects(expression.mgl_jsonExpressionObject, @[@"properties"]);
-        XCTAssertEqualObjects([NSExpression expressionWithFormat:@"$featureProperties"].mgl_jsonExpressionObject, @[@"properties"]);
+        XCTAssertEqualObjects([NSExpression expressionWithFormat:@"$featureAttributes"].mgl_jsonExpressionObject, @[@"properties"]);
         XCTAssertEqualObjects([NSExpression expressionWithMGLJSONObject:@[@"properties"]], expression);
     }
     {
@@ -867,7 +867,7 @@ using namespace std::string_literals;
     }
     {
         NSExpression *expression = [NSExpression expressionForFunction:@"mgl_does:have:"
-                                                             arguments:@[[NSExpression expressionForVariable:@"featureProperties"],
+                                                             arguments:@[[NSExpression expressionForVariable:@"featureAttributes"],
                                                                          [NSExpression expressionForConstantValue:@"x"]]];
         NSArray *jsonExpression = @[@"has", @"x", @[@"properties"]];
         XCTAssertEqualObjects(expression.mgl_jsonExpressionObject, jsonExpression);